From the very first time...
It worked!

So this commit preserves forever the way that works.

Next I will test which of my 2 changes actually made istio work:
a. Not using the KUD job to deploy k8s on the VMs
or
b. Only ading the global suffix to the coredns configmap instead of a
total replace according to the istio website itself.