This commit is to confirm that to deploy Istio
KUD jobs for installing on VMs cannot be used.

There is some sort of incompatibility with the specific
deployment made by the job. Granted, it could be that I'm
also disabling extra addons when I do the manual aio.sh
install in the VMs (I only decided to do it because QAT
plugin fails to install).